core/json: Rename json_cache.rs to cache.rs
