+++ core/json/cache.rs
